Refactor pipeline into stages and modular crawler
