+++ backend/pipeline/__init__.py
@@ -6,12 +6,14 @@ from . import step_3_classify_reviews
 from . import step_4_link_movies
 from . import step_5_generate_sentiment
 from . import step_6_enrich_metadata
+from . import step_3b_llm_validation
 
 __all__ = [
     "step_1_fetch_links",
     "step_2_parse_posts",
     "step_3_classify_reviews",
     "step_4_link_movies",
+    "step_3b_llm_validation",
     "step_5_generate_sentiment",
     "step_6_enrich_metadata",
 ]
