+++ backend/pipeline/step_1_fetch_links.py
@@ -5,17 +5,17 @@ from db.store_review import store_blog_post_urls
 from utils import StepLogger
 
 
-async def fetch_links(start_page: int = 1, end_page: int = 279) -> list[str]:
+async def fetch_links(start_page: int = 1, end_page: int = 279, reviewer: str = "baradwajrangan") -> list[str]:
     """Fetch blog post links and persist them."""
     step_logger = StepLogger("step_1_fetch_links")
-    links = await get_post_links_async(start_page, end_page)
+    links = await get_post_links_async(start_page, end_page, reviewer)
     step_logger.metrics["input_count"] = len(links)
     step_logger.logger.info("Discovered %s new links", len(links))
 
     for link in links:
         step_logger.metrics["processed_count"] += 1
         try:
-            store_blog_post_urls({"link": link[2], "blog_title": "TBD"})
+            store_blog_post_urls({"link": link[2], "blog_title": "TBD", "reviewer": reviewer})
             step_logger.metrics["saved_count"] += 1
         except Exception as e:
             step_logger.metrics["failed_count"] += 1
